[Python] fixed: Python plugin will work for the firs time, but when Plugin is run again after DLL is unloaded from memory plugin will crash and UI will be locked
- This commit is specific to Xbox. In XBMC Python is handled by Operating System so XBMC DllLoader is not used. Because of this XBMC will create and allocate XBMC types only once. Hovever this won't work on Xbox because we
are loading and unloading Python DLL each time when plugin is launched. This means that memory locations of Python types won't be same, they are always changing when Plugin is launched. In order to fix this XBMC types will always be recreated on Python launch. This is how it worked before on XBMC4Xbox.